Moved ExtraDataMenu component
